Works!

+++ src/com/github/dreamrec/AveragingFilter.java
@@ -3,21 +3,21 @@ package com.github.dreamrec;
 /**
  *
  */
-public class AveragingFilter extends AbstractFilter<Short> {
+public class AveragingFilter extends AbstractFilter<Integer> {
 
-    public AveragingFilter(Filter<Short> inputData, int divider) {
+    public AveragingFilter(Filter<Integer> inputData, int divider) {
         super(inputData);
         this.divider = divider;
     }
 
     @Override
-    protected Short doFilter(int index) {
+    protected Integer doFilter(int index) {
         int sum = 0;
         int incomingDataIndex = index*divider;
         for (int i = 0; i < divider; i++) {
             sum += Math.abs(inputData.get(incomingDataIndex+i));
 
         }
-        return (short)(sum*100/divider);
+        return sum*100/divider;
     }
 }
