+++ src/com/github/dreamrec/MainWindow.java
@@ -57,7 +57,7 @@ public class MainWindow extends JFrame {
         mainPanel.add(ch2DataView);
         ch2DataView.getComponentModel().centreX();
 
-        Filter<Short> slowDreamView = new AveragingFilter(new FirstDerivativeAbsFilter(model.getEyeDataList()), Model.DIVIDER);
+        Filter<Integer> slowDreamView = new AveragingFilter(new FirstDerivativeAbsFilter(model.getEyeDataList()), Model.DIVIDER);
         mainPanel.add(Factory.getGComponentView(model, controller,slowDreamView));
 
         add(mainPanel, BorderLayout.CENTER);
